Warp the mouse to a newly created window.
The warp happens only if the window wasn't created hidden and
if the mouse isn't already inside a new rectangle.
The later is perhaps a questionable choice and should be
changed to only warp if the mouse isn't on either of the four
corners, i.e.  if the user didn't manually shape the new
window.