Revert all followfocus
The experiment (mostly) failed.  What we actually want is
passing the mouse event that caused a focus change through to
the application running in a newly focused window.
For example b1 click on a blurred samterm would would raise it
and give it input focus while also (possibly) changing the
cursor position samterm itself, without requiring a second
click.